+++ backend/src/baserow/contrib/database/table/exceptions.py
@@ -2,6 +2,18 @@ class TableDoesNotExist(Exception):
     """Raised when trying to get a table that doesn't exist."""
 
 
+class TableNotInDatabase(Exception):
+    """Raised when a provided table does not belong to a database."""
+
+    def __init__(self, table_id=None, *args, **kwargs):
+        self.table_id = table_id
+        super().__init__(
+            f"The table {table_id} does not belong to the database.",
+            *args,
+            **kwargs,
+        )
+
+
 class InvalidInitialTableData(Exception):
     """Raised when the provided initial table data does not contain a column or row."""