+++ web-frontend/modules/database/components/sidebar/SidebarItem.vue
@@ -1,16 +1,19 @@
 <template>
   <li class="tree__sub" :class="{ active: table._.selected }">
-    <a class="tree__sub-link" @click.prevent="selectTable(database, table)">
-      <Editable
-        ref="rename"
-        :value="table.name"
-        @change="renameTable(database, table, $event)"
-      ></Editable>
+    <a class="tree__sub-link" @click="selectTable(database, table)">
+      <div>
+        <Editable
+          ref="rename"
+          :value="table.name"
+          @change="renameTable(database, table, $event)"
+        ></Editable>
+      </div>
     </a>
     <a
       v-show="!database._.loading"
       class="tree__options"
       @click="$refs.context.toggle($event.currentTarget, 'bottom', 'right', 0)"
+      @mousedown.stop
     >
       <i class="fas fa-ellipsis-v"></i>
     </a>